# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: cdCTF API
  version: 0.1.0
  description: cdCTF API
servers:
- url: /api
  description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUserProfile
      tags: [profile]
      summary: Get user profile
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateUserProfile
      tags: [profile]
      summary: Update user profile
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"


    delete:
      operationId: deleteAccount
      tags: [users]
      summary: Delete an account — the caller's own, or any as an admin
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /admin/competitions/{id}/users/{userId}:
    post:
      operationId: adminAddCompetitionUser
      tags: [admin]
      summary: Add user to private competition
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
      - in: path
        name: userId
        required: true
        schema:
          type: integer
      responses:
        "200":
          description: User added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /users/me/dashboard:
    get:
      operationId: getMyDashboard
      tags: [users]
      summary: The caller's progress summary and rank
      responses:
        "200":
          description: Dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users/me/profile:
    get:
      operationId: getMyProfile
      tags: [users]
      summary: The caller's own profile, including their email
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users/{id}/avatar:
    post:
      operationId: uploadAvatar
      tags: [users]
      summary: Upload an avatar (2 MB max, images only)
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                avatar:
                  type: string
                  format: binary
              required: [avatar]
      responses:
        "200":
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          description: File too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # COMPETITIONS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # Auth
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        nickname:
          type: string
        email:
          type: string
        avatarUrl:
          type: string
          nullable: true
        points:
          type: integer
        rank:
          type: integer
        isBlocked:
          type: boolean
        solvedCtf:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              category:
                type: string
              points:
                type: integer
              solvedAt:
                type: string
                format: date-time
            required: [id, name, category, points, solvedAt]
        completedLessons:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              title:
                type: string
              points:
                type: integer
              completedAt:
                type: string
                format: date-time
            required: [id, title, points, completedAt]
        competitionHistory:
          type: array
          items:
            type: object
            properties:
              competitionId:
                type: integer
              competitionName:
                type: string
              rank:
                type: integer
              points:
                type: integer
            required: [competitionId, competitionName, rank, points]
        titles:
          type: array
          items:
            $ref: "#/components/schemas/Title"
        createdAt:
          type: string
          format: date-time
      required: [id, nickname, email, points, rank, isBlocked, solvedCtf, completedLessons, competitionHistory, titles, createdAt]
    UpdateProfileBody:
      type: object
      properties:
        nickname:
          type: string
          minLength: 3
          maxLength: 32

        avatarUrl:
          type: string
          nullable: true
        points:
          type: integer
          description: Admin only.
        role:
          type: string
          enum:
          - user
          - author
          - moderator
          - admin
          description: Admin only.
        isBlocked:
          type: boolean
          description: Admin only.
        email:
          type: string
          format: email
          description: Admin only.
        emailVerified:
          type: boolean
          description: Admin only.
      description: Fields this endpoint accepts. Which of them a caller may actually set depends on their role (columnPermissions in rbac.ts); anything they may not set is dropped, and a request left with nothing to update is rejected.
    AvatarResponse:
      type: object
      properties:
        avatarUrl:
          type: string
      required: [avatarUrl]
    Title:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        points:
          type: integer
          description: Bonus points for earning this title
        earnedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, name, category, points]

    # Admin
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    DashboardResponse:
      type: object
      properties:
        user:
          type: object
          properties:
            id:
              type: integer
            nickname:
              type: string
            points:
              type: integer
            rank:
              type: integer
              description: 0 for accounts excluded from scoring.
          required: [id, nickname, points, rank]
        progress:
          type: object
          properties:
            solvedCtfCount:
              type: integer
            completedLessonCount:
              type: integer
            titleCount:
              type: integer
          required: [solvedCtfCount, completedLessonCount, titleCount]
        recent:
          type: object
          properties:
            solvedCtf:
              type: array
              items:
                type: object
                properties:
                  ctfId:
                    type: integer
                  solvedAt:
                    type: string
                    format: date-time
                    nullable: true
            completedLessons:
              type: array
              items:
                type: object
                properties:
                  lessonId:
                    type: integer
                  completedAt:
                    type: string
                    format: date-time
                    nullable: true
        titles:
          type: array
          items:
            $ref: "#/components/schemas/Title"
      required: [user, progress, recent, titles]
...
  responses:
    Unauthorized:
      description: Not signed in, or the token is invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Signed in, but not allowed to do this
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...